% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{English_lu}
\alias{English_lu}
\title{Diceware English Word table}
\format{A data frame with 7776 rows and 2 variables:
\describe{
  \item{roll}{5 numbers 1-6}
  \item{word}{words associated with the rolls}
  ...
}}
\source{
\url{http://world.std.com/~reinhold/diceware.wordlist.asc}
}
\usage{
English_lu
}
\description{
A dataset containing the English words for diceware passphrases
}
\keyword{datasets}

