% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{phrase}
\alias{phrase}
\title{Generate a passphrase}
\usage{
phrase(nwords)
}
\arguments{
\item{nwords}{The number of words in your passphrase}
}
\value{
A character string with the dice rolls (from Random.org) and the diceware lookup words
}
\description{
Generate a passphrase
}

