% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{wordlu}
\alias{wordlu}
\title{Find the word associated with a roll}
\usage{
wordlu(roll, lu = dicewaRe:::English_lu)
}
\arguments{
\item{roll}{A number that will be found in the "roll" column of the "lu" object.}

\item{lu}{A data frame with a "roll" and a "word" column}
}
\value{
A word from a lookup table
}
\description{
Find the word associated with a roll
}

